feat: eliminate all index signatures with --strict
